third_party/blink/web_tests/external/wpt/mediacapture-streams/MediaStream-removetrack.https-expected.txt

This is a testharness.js-based test.
[FAIL] Test that removal from a MediaStream fires ended on media elements (video first)
  assert_equals: audio element ended because no more audio tracks expected true but got false
[FAIL] Test that removal from a MediaStream fires ended on media elements (audio first)
  assert_equals: audio element ended because no more audio tracks expected true but got false
Harness: the test ran to completion.
